requirements.txt
Django==4.2.5
mysqlclient==2.1.1
google-generativeai
PyPDF2==2.0.0
pymupdf